refactor(viewer): one collapsible mode-aware Sun & Sky panel; OSM base map + shadow receiving
Feedback round on the sky/lighting feature:

- Merge SolarPanel + EnvironmentPanel into a single collapsible Sun & Sky
  panel (one toolbar button). Mode-aware: standalone shows lighting preset +
  exposure (they shape the WebGPU shading); in the Cesium world context the
  GLB is unlit and lit by sun/atmosphere, so those controls hide and Sky
  drives the atmosphere instead.
- Panel anchors below the ViewCube instead of covering it.
- Sweep animation moves to useSolarSweep at the viewport level so closing
  the panel doesn't stop a running sweep; shared time helpers in
  lib/solar-time.ts.
- OSM Buildings mode keeps the globe with the world-imagery base map —
  buildings sit ON TOP of the satellite image instead of replacing it, and
  the globe (ShadowMode.RECEIVE_ONLY) is what makes building/model shadows
  visible during a sun study.
- Night preset ships an explicit dark sky (the altitude-derived palette read
  the high moon as midday); derived ground tones neutralised to grey since
  top-down cameras fill the screen with the below-horizon colour.
- Panel warns when the projected CRS can't be resolved (no site → no real
  sun position).
- Drop the envSunFollowsSolar toggle: an active sun study simply drives the
  sun.
